move most of the contextual menu logic to PlaygroundFilePicker
