+++ packages/playground/website/src/php-playground/components/Layout.tsx
@@ -8,8 +8,9 @@ import styles from './layout.module.css';
 import { Controls } from './Controls';
 import { EditorHost } from './EditorHost';
 import { HelpModal } from './HelpModal';
-import PlaygroundFilePickerTree from './PlaygroundFilePickerTree';
-import type { FileNode } from '@wp-playground/components';
+import PlaygroundFilePickerTree, {
+	type PlaygroundFilePickerTreeRef,
+} from './PlaygroundFilePickerTree';
 import { PlaygroundManager } from './PlaygroundManager';
 import { Terminal } from './Terminal';
 import { useAppDispatch, useAppSelector } from '../hooks';
@@ -45,15 +46,7 @@ const dirnameSafe = (path: string) => {
 
 // removed legacy prompt-based helpers in favor of inline rename flow
 
-const isValidNameSegment = (name: string) => {
-	if (!name) {
-		return false;
-	}
-	if (name === '.' || name === '..') {
-		return false;
-	}
-	return !/[\\/]/.test(name);
-};
+// validation handled in tree component
 
 const isProbablyTextBuffer = (buffer: Uint8Array) => {
 	// Fast null-byte check
@@ -85,11 +78,11 @@ export const Layout = () => {
 	const [isTerminalCollapsed, setTerminalCollapsed] = useState(false);
 	const [terminalResizeToken, setTerminalResizeToken] = useState(0);
 	const terminalPanelRef = useRef<ImperativePanelHandle | null>(null);
+	const treeRef = useRef<PlaygroundFilePickerTreeRef | null>(null);
 	const playgroundClient = useAppSelector((state) => state.playground.client);
 	const bootStatus = useAppSelector((state) => state.playground.bootStatus);
 	const currentPath = useAppSelector((state) => state.playground.currentPath);
 	const [previewUrl, setPreviewUrl] = useState('');
-	const navigationClientRef = useRef<PlaygroundClient | null>(null);
 	const [renamingPath, setRenamingPath] = useState<string | null>(null);
 	const [selectedDirPath, setSelectedDirPath] = useState<string | null>(
 		DEFAULT_WORKSPACE_DIR
@@ -97,52 +90,8 @@ export const Layout = () => {
 	const [forceSelectedPath, setForceSelectedPath] = useState<string | null>(
 		null
 	);
-	const pendingCreateRef = useRef<{
-		type: 'file' | 'folder';
-		tempPath: string;
-	} | null>(null);
-	const [contextMenu, setContextMenu] = useState<{
-		path: string;
-		type: 'file' | 'folder';
-		x: number;
-		y: number;
-	} | null>(null);
 
 	const dispatch = useAppDispatch();
-
-	// getBaseName not used in this component
-
-	const findAvailableName = useCallback(
-		async (client: PlaygroundClient, baseDir: string, baseName: string) => {
-			let name = baseName;
-			let counter = 0;
-			const splitExt = (n: string) => {
-				const dot = n.lastIndexOf('.');
-				if (dot > 0) {
-					return { stem: n.slice(0, dot), ext: n.slice(dot) };
-				}
-				return { stem: n, ext: '' };
-			};
-			const prefix = baseDir === '/' ? '' : baseDir;
-			while (
-				(await client
-					.fileExists(`${prefix}/${name}`)
-					.catch(() => false)) ||
-				(await client.isDir(`${prefix}/${name}`).catch(() => false))
-			) {
-				counter += 1;
-				const { stem, ext } = splitExt(baseName);
-				name = `${stem} (${counter})${ext}`;
-			}
-			return name;
-		},
-		[]
-	);
-
-	const ensureParentExpandedPath = useCallback((path: string) => {
-		const parent = dirnameSafe(path);
-		setSelectedDirPath(parent);
-	}, []);
 
 	useEffect(() => {
 		const previousTitle = document.title;
@@ -151,26 +100,6 @@ export const Layout = () => {
 			document.title = previousTitle;
 		};
 	}, []);
-
-	useEffect(() => {
-		if (!contextMenu) {
-			return;
-		}
-		const handleClose = () => setContextMenu(null);
-		const handleKey = (event: KeyboardEvent) => {
-			if (event.key === 'Escape') {
-				handleClose();
-			}
-		};
-		window.addEventListener('click', handleClose);
-		window.addEventListener('contextmenu', handleClose);
-		window.addEventListener('keydown', handleKey);
-		return () => {
-			window.removeEventListener('click', handleClose);
-			window.removeEventListener('contextmenu', handleClose);
-			window.removeEventListener('keydown', handleKey);
-		};
-	}, [contextMenu]);
 
 	useEffect(() => {
 		if (!isHelpOpen) {
@@ -187,11 +116,6 @@ export const Layout = () => {
 		if (!playgroundClient) {
 			return;
 		}
-		// Avoid registering multiple onNavigation handlers for the same client
-		if (navigationClientRef.current === playgroundClient) {
-			return;
-		}
-		navigationClientRef.current = playgroundClient;
 		let cancelled = false;
 		(async () => {
 			try {
@@ -216,274 +140,9 @@ export const Layout = () => {
 	const handleOpenHelp = useCallback(() => setHelpOpen(true), []);
 	const handleCloseHelp = useCallback(() => setHelpOpen(false), []);
 
-	const handleCreateFile = useCallback(async () => {
-		if (!playgroundClient) {
-			return;
-		}
-		const baseDir = selectedDirPath
-			? normalizeFsPath(selectedDirPath)
-			: currentPath
-			? dirnameSafe(currentPath)
-			: DEFAULT_WORKSPACE_DIR;
-		const normalizedBase = normalizeFsPath(baseDir);
-		try {
-			const name = await findAvailableName(
-				playgroundClient,
-				normalizedBase,
-				'untitled.php'
-			);
-			const tempPath = `${
-				normalizedBase === '/' ? '' : normalizedBase
-			}/${name}`;
-			await playgroundClient.writeFile(tempPath, '');
-			pendingCreateRef.current = { type: 'file', tempPath };
-			setRenamingPath(tempPath);
-			setForceSelectedPath(tempPath);
-			ensureParentExpandedPath(tempPath);
-			setSelectedDirPath(dirnameSafe(tempPath));
-			setContextMenu(null);
-		} catch (e) {
-			void e;
-		}
-	}, [
-		playgroundClient,
-		currentPath,
-		selectedDirPath,
-		findAvailableName,
-		ensureParentExpandedPath,
-	]);
-
-	const handleCreateDirectory = useCallback(async () => {
-		if (!playgroundClient) {
-			return;
-		}
-		const baseDir = selectedDirPath
-			? normalizeFsPath(selectedDirPath)
-			: currentPath
-			? dirnameSafe(currentPath)
-			: DEFAULT_WORKSPACE_DIR;
-		const normalizedBase = normalizeFsPath(baseDir);
-		try {
-			const name = await findAvailableName(
-				playgroundClient,
-				normalizedBase,
-				'New Folder'
-			);
-			const tempPath = `${
-				normalizedBase === '/' ? '' : normalizedBase
-			}/${name}`;
-			await playgroundClient.mkdir(tempPath);
-			pendingCreateRef.current = { type: 'folder', tempPath };
-			setRenamingPath(tempPath);
-			setForceSelectedPath(tempPath);
-			ensureParentExpandedPath(tempPath);
-			setSelectedDirPath(tempPath);
-			setContextMenu(null);
-		} catch (e) {
-			void e;
-		}
-	}, [
-		playgroundClient,
-		currentPath,
-		selectedDirPath,
-		findAvailableName,
-		ensureParentExpandedPath,
-	]);
-
-	const handleRenameSubmit = useCallback(
-		async (path: string, newName: string) => {
-			if (!playgroundClient) {
-				return;
-			}
-			const pending = pendingCreateRef.current;
-			const isPending = pending?.tempPath === path;
-			const parent = dirnameSafe(path);
-			const sanitized = (newName || '').trim();
-			if (!isValidNameSegment(sanitized)) {
-				if (isPending) {
-					try {
-						if (pending.type === 'folder') {
-							await playgroundClient.rmdir(path, {
-								recursive: true,
-							} as any);
-						} else {
-							await playgroundClient.unlink(path);
-						}
-					} catch (e) {
-						void e;
-					}
-					pendingCreateRef.current = null;
-				}
-				setRenamingPath(isPending ? null : path);
-				return;
-			}
-
-			const composePath = (name: string) =>
-				parent === '/' ? `/${name}` : `${parent}/${name}`;
-			const finalName = sanitized;
-			let candidate = composePath(finalName);
-			const candidateNormalized = normalizeFsPath(candidate);
-			if (candidateNormalized === path) {
-				setRenamingPath(null);
-				if (isPending) {
-					pendingCreateRef.current = null;
-				}
-				setForceSelectedPath(candidateNormalized);
-				return;
-			}
-
-			const exists = await playgroundClient
-				.fileExists(candidateNormalized)
-				.catch(() => false);
-			const existsDir = await playgroundClient
-				.isDir(candidateNormalized)
-				.catch(() => false);
-			if ((exists || existsDir) && candidateNormalized !== path) {
-				if (isPending) {
-					try {
-						const unique = await findAvailableName(
-							playgroundClient,
-							parent === '/' ? '/' : parent,
-							finalName
-						);
-						candidate = composePath(unique);
-					} catch (e) {
-						void e;
-					}
-				} else {
-					setRenamingPath(path);
-					return;
-				}
-			}
-
-			let candidateIsDir = pending?.type === 'folder';
-
-			try {
-				await playgroundClient.mv(path, candidate as any);
-				if (pending?.type === 'file') {
-					try {
-						const newContent =
-							await playgroundClient.readFileAsText(candidate);
-						dispatch(setCode(newContent));
-						dispatch(setCurrentPath(candidate));
-					} catch (e) {
-						void e;
-					}
-				} else if (!pending) {
-					const isDir = await playgroundClient
-						.isDir(candidate as any)
-						.catch(() => false);
-					candidateIsDir = isDir;
-					if (!isDir) {
-						try {
-							const newContent =
-								await playgroundClient.readFileAsText(
-									candidate
-								);
-							dispatch(setCode(newContent));
-							dispatch(setCurrentPath(candidate));
-						} catch {
-							// Ignore failure
-						}
-					} else if (currentPath === path) {
-						dispatch(setCurrentPath(candidate));
-					}
-				}
-				setForceSelectedPath(candidate);
-				setSelectedDirPath(
-					candidateIsDir ? candidate : dirnameSafe(candidate)
-				);
-			} catch {
-				if (isPending) {
-					try {
-						if (pending?.type === 'folder') {
-							await playgroundClient.rmdir(path, {
-								recursive: true,
-							} as any);
-						} else {
-							await playgroundClient.unlink(path);
-						}
-					} catch {
-						// Ignore failure
-					}
-				}
-			} finally {
-				pendingCreateRef.current = null;
-				setRenamingPath(null);
-			}
-		},
-		[playgroundClient, findAvailableName, dispatch, currentPath]
-	);
-
-	const handleRenameCancel = useCallback(
-		async (path: string) => {
-			const pending = pendingCreateRef.current;
-			setContextMenu(null);
-			if (!playgroundClient || pending?.tempPath !== path) {
-				setRenamingPath(null);
-				return;
-			}
-			try {
-				if (pending.type === 'folder') {
-					await playgroundClient.rmdir(path, {
-						recursive: true,
-					} as any);
-				} else {
-					await playgroundClient.unlink(path);
-				}
-			} catch (e) {
-				void e;
-			}
-			pendingCreateRef.current = null;
-			setRenamingPath(null);
-			setForceSelectedPath(null);
-			setSelectedDirPath((prev) => {
-				if (!prev) {
-					return prev;
-				}
-				const normalized = normalizeFsPath(path);
-				if (prev === normalized || prev.startsWith(`${normalized}/`)) {
-					return dirnameSafe(normalized);
-				}
-				return prev;
-			});
-		},
-		[playgroundClient]
-	);
-
-	const handleContextMenuRename = useCallback(
-		(path: string) => {
-			pendingCreateRef.current = null;
-			setContextMenu(null);
-			setRenamingPath(path);
-			setForceSelectedPath(path);
-			setSelectedDirPath(dirnameSafe(path));
-			ensureParentExpandedPath(path);
-		},
-		[ensureParentExpandedPath]
-	);
-
-	const handleDeletePath = useCallback(
-		async (targetPath: string, type: 'file' | 'folder') => {
-			if (!playgroundClient) {
-				return;
-			}
-			const normalized = normalizeFsPath(targetPath);
-			setContextMenu(null);
-			try {
-				if (type === 'folder') {
-					await playgroundClient.rmdir(normalized, {
-						recursive: true,
-					} as any);
-				} else {
-					await playgroundClient.unlink(normalized);
-				}
-			} catch {
-				// Ignore failure
-			}
-			if (pendingCreateRef.current?.tempPath === normalized) {
-				pendingCreateRef.current = null;
-			}
+	const handleAfterDelete = useCallback(
+		(path: string, type: 'file' | 'folder') => {
+			const normalized = normalizeFsPath(path);
 			if (
 				currentPath &&
 				(currentPath === normalized ||
@@ -502,52 +161,27 @@ export const Layout = () => {
 			setRenamingPath(null);
 			setForceSelectedPath(dirnameSafe(normalized));
 		},
-		[playgroundClient, currentPath, selectedDirPath, dispatch]
+		[currentPath, selectedDirPath, dispatch]
 	);
 
-	const handleDownloadPath = useCallback(
-		async (targetPath: string) => {
-			if (!playgroundClient) {
-				return;
-			}
-			setContextMenu(null);
+	const handleAfterRename = useCallback(
+		async (oldPath: string, newPath: string, type: 'file' | 'folder') => {
+			if (!playgroundClient) return;
 			try {
-				const content = await playgroundClient.readFileAsText(
-					targetPath
-				);
-				const blob = new Blob([content], {
-					type: 'text/plain;charset=utf-8',
-				});
-				const url = URL.createObjectURL(blob);
-				const link = document.createElement('a');
-				link.href = url;
-				link.download =
-					normalizeFsPath(targetPath).split('/').pop() || 'template';
-				document.body.appendChild(link);
-				link.click();
-				link.remove();
-				URL.revokeObjectURL(url);
-			} catch {
-				// Ignore failure
+				if (type === 'file') {
+					const newContent = await playgroundClient.readFileAsText(
+						newPath
+					);
+					dispatch(setCode(newContent));
+					dispatch(setCurrentPath(newPath));
+				} else if (currentPath === oldPath) {
+					dispatch(setCurrentPath(newPath));
+				}
+			} catch (e) {
+				void e;
 			}
 		},
-		[playgroundClient]
-	);
-
-	const handleTreeContextMenu = useCallback(
-		(event: React.MouseEvent, node: FileNode, path: string) => {
-			event.preventDefault();
-			event.stopPropagation();
-			const absolute = normalizeFsPath(path);
-			setRenamingPath(null);
-			setContextMenu({
-				path: absolute,
-				type: node.type,
-				x: event.clientX,
-				y: event.clientY,
-			});
-		},
-		[renamingPath]
+		[playgroundClient, currentPath, dispatch]
 	);
 
 	const treeInitialPath = normalizeFsPath(
@@ -556,19 +190,6 @@ export const Layout = () => {
 			selectedDirPath ??
 			(currentPath ? dirnameSafe(currentPath) : DEFAULT_WORKSPACE_DIR)
 	);
-
-	const contextMenuStyle = contextMenu
-		? {
-				left:
-					typeof window === 'undefined'
-						? contextMenu.x
-						: Math.min(contextMenu.x, window.innerWidth - 200),
-				top:
-					typeof window === 'undefined'
-						? contextMenu.y
-						: Math.min(contextMenu.y, window.innerHeight - 140),
-		  }
-		: undefined;
 
 	return (
 		<div id="php-playground-react-root" className={styles.root}>
@@ -594,7 +215,9 @@ export const Layout = () => {
 												className={
 													styles.fileExplorerButton
 												}
-												onClick={handleCreateFile}
+												onClick={() =>
+													treeRef.current?.createFile()
+												}
 												title="Create new file"
 												disabled={!playgroundClient}
 											>
@@ -604,7 +227,9 @@ export const Layout = () => {
 												className={
 													styles.fileExplorerButton
 												}
-												onClick={handleCreateDirectory}
+												onClick={() =>
+													treeRef.current?.createFolder()
+												}
 												title="Create new folder"
 												disabled={!playgroundClient}
 											>
@@ -614,6 +239,7 @@ export const Layout = () => {
 									</div>
 									<div className={styles.fileExplorerTree}>
 										<PlaygroundFilePickerTree
+											ref={treeRef}
 											playgroundClient={playgroundClient}
 											root="/"
 											initialPath={treeInitialPath}
@@ -624,7 +250,6 @@ export const Layout = () => {
 												'/request',
 											]}
 											onSelect={async (path) => {
-												setContextMenu(null);
 												setForceSelectedPath(null);
 												if (
 													await playgroundClient.isDir(
@@ -721,11 +346,8 @@ export const Layout = () => {
 												}
 											}}
 											renamingPath={renamingPath}
-											onRename={handleRenameSubmit}
-											onRenameCancel={handleRenameCancel}
-											onContextMenu={
-												handleTreeContextMenu
-											}
+											onAfterDelete={handleAfterDelete}
+											onAfterRename={handleAfterRename}
 										/>
 									</div>
 								</div>
@@ -817,42 +439,6 @@ export const Layout = () => {
 				</Panel>
 			</PanelGroup>
 			<HelpModal isOpen={isHelpOpen} onRequestClose={handleCloseHelp} />
-			{contextMenu && contextMenuStyle && (
-				<div
-					className={styles.contextMenu}
-					style={contextMenuStyle}
-					onClick={(event) => event.stopPropagation()}
-					onContextMenu={(event) => event.preventDefault()}
-				>
-					<button
-						className={styles.contextMenuButton}
-						onClick={() =>
-							handleContextMenuRename(contextMenu.path)
-						}
-					>
-						Rename
-					</button>
-					<button
-						className={clsx(
-							styles.contextMenuButton,
-							styles.contextMenuButtonDanger
-						)}
-						onClick={() =>
-							handleDeletePath(contextMenu.path, contextMenu.type)
-						}
-					>
-						Delete
-					</button>
-					{contextMenu.type === 'file' && (
-						<button
-							className={styles.contextMenuButton}
-							onClick={() => handleDownloadPath(contextMenu.path)}
-						>
-							Download
-						</button>
-					)}
-				</div>
-			)}
 		</div>
 	);
 };
